requirements.txt
Django==2.1.7
django-bootstrap3==11.0.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-filter==2.1.0
djangorestframework==3.9.1